dcp-815 separate link_set.py
